+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to
 - Save cluster settings on deployment so they can reused automatically in
   subsequent runs of `firehpc conf` and `firehpc restore` (#7).
 - Report cluster settings in `firehpc status`.
+- Introduce `firehpc update` command to change cluster settings.
 - load:
   - Submit jobs randomly in existing QOS and partitions.
   - Submit jobs of various sizes, with a power of 2 number (1, 2, 4, 8…) of